+++ internal/server/crafting.go
@@ -628,7 +628,7 @@ func (h *hub) tossItem(players map[int32]*tracked, t *tracked, st invStack) {
 		it.ench = st.ench
 		it.mapID = st.mapID
 		it.pats = st.pats
-		it.trimMat, it.trimPat = st.trimMat, st.trimPat
+		it.trimMat, it.trimPat, it.color = st.trimMat, st.trimPat, st.color
 		it.bookID = st.bookID
 		it.boxID, it.hiveID = st.boxID, st.hiveID
 		it.bundleID, it.potion, it.repairCost, it.instrument, it.name, it.lode = st.bundleID, st.potion, st.repairCost, st.instrument, st.name, st.lode